Coffee Shop Application
The Coffee Shop Application is here to help with the normal work of the cafe. The application records all orders that are not yet ready and reminds the employees of them by visualizing a picture of the category, the product's name and its price. It also calculates the approximate time that will be required for the execution of the orders. As an additional functionality, statistics are kept, on which employees can process how many orders at that moment.  At this stage, the functionality of the application is small but later it can be expanded upon.